Separate classes into files to comply with CodeFactor
