+++ Files/Files.csproj
@@ -184,7 +184,12 @@
       <DependentUpon>RestartDialog.xaml</DependentUpon>
     </Compile>
     <Compile Include="Helpers\RegistryHelper.cs" />
+    <Compile Include="UserControls\Selection\ExtendPreviousItemSelectionStrategy.cs" />
+    <Compile Include="UserControls\Selection\GenericSelectedItems.cs" />
+    <Compile Include="UserControls\Selection\IgnorePreviousItemSelectionStrategy.cs" />
+    <Compile Include="UserControls\Selection\ISelectedItems.cs" />
     <Compile Include="UserControls\Selection\ItemSelectionStrategy.cs" />
+    <Compile Include="UserControls\Selection\RawSelectedItems.cs" />
     <Compile Include="View Models\FolderSettingsViewModel.cs" />
     <Compile Include="View Models\Properties\FileProperty.cs" />
     <Compile Include="Extensions\EnumerableExtensions.cs" />
