+++ Files/UserControls/Selection/ISelectedItems.cs
@@ -1,0 +1,13 @@
+﻿namespace Files.UserControls.Selection
+{
+    public interface ISelectedItems
+    {
+        void Add(object item);
+
+        void Clear();
+
+        bool Contains(object item);
+
+        void Remove(object item);
+    }
+}
